deleted:    blockchain/contracts/APIConsumer.sol 	modified:   blockchain/contracts/AminoChainAuthenticator.sol 	new file:   blockchain/contracts/AminoChainMarketplace.sol 	modified:   blockchain/contracts/DonationNFT.sol 	deleted:    blockchain/contracts/KeepersCounter.sol 	deleted:    blockchain/contracts/PriceConsumerV3.sol 	deleted:    blockchain/contracts/RandomNumberConsumerV2.sol 	renamed:    blockchain/contracts/IAminoChainAuthenticator.sol -> blockchain/contracts/interfaces/IAminoChainAuthenticator.sol 	renamed:    blockchain/contracts/IAminoChainMarketplace.sol -> blockchain/contracts/interfaces/IAminoChainMarketplace.sol 	renamed:    blockchain/contracts/IDonationNFT.sol -> blockchain/contracts/interfaces/IDonationNFT.sol 	renamed:    blockchain/contracts/AminoChainLibrary.sol -> blockchain/contracts/libraries/AminoChainLibrary.sol 	new file:   blockchain/contracts/mocks/MockERC20.sol 	renamed:    blockchain/contracts/test/MockNFT.sol -> blockchain/contracts/mocks/MockNFT.sol 	renamed:    blockchain/contracts/test/MockOracle.sol -> blockchain/contracts/mocks/MockOracle.sol 	renamed:    blockchain/contracts/test/MockV3Aggregator.sol -> blockchain/contracts/mocks/MockV3Aggregator.sol 	renamed:    blockchain/contracts/test/Token.sol -> blockchain/contracts/mocks/Token.sol 	renamed:    blockchain/contracts/test/VRFCoordinatorV2Mock.sol -> blockchain/contracts/mocks/VRFCoordinatorV2Mock.sol 	deleted:    blockchain/contracts/test/MockAminoChainMarketplace.sol 	deleted:    blockchain/contracts/test/fuzzing/KeepersCounterEchidnaTest.sol 	deleted:    blockchain/contracts/test/fuzzing/config.yaml 	modified:   blockchain/hardhat.config.ts 	modified:   blockchain/helper-hardhat-config.ts 	modified:   blockchain/typechain/@chainlink/contracts/src/v0.8/index.ts 	modified:   blockchain/typechain/@chainlink/contracts/src/v0.8/interfaces/index.ts 	new file:   blockchain/typechain/@openzeppelin/contracts/token/ERC721/extensions/ERC721Enumerable.ts 	new file:   blockchain/typechain/@openzeppelin/contracts/token/ERC721/extensions/IERC721Enumerable.ts 	new file:   blockchain/typechain/contracts/AminoChainMarketplace.ts 	new file:   blockchain/typechain/contracts/DonationNFT.ts 	modified:   blockchain/typechain/contracts/index.ts 	new file:   blockchain/typechain/contracts/interfaces/IAminoChainAuthenticator.ts 	new file:   blockchain/typechain/contracts/interfaces/IAminoChainMarketplace.ts 	new file:   blockchain/typechain/contracts/interfaces/IDonationNFT.ts 	new file:   blockchain/typechain/contracts/interfaces/index.ts 	new file:   blockchain/typechain/contracts/mocks/MockAminoChainMarketplace.ts 	new file:   blockchain/typechain/contracts/mocks/MockERC20.ts 	new file:   blockchain/typechain/contracts/mocks/MockNFT copy.sol/MockNFT.ts 	new file:   blockchain/typechain/contracts/mocks/MockNFT copy.sol/index.ts 	new file:   blockchain/typechain/contracts/mocks/MockNFT.ts 	new file:   blockchain/typechain/contracts/mocks/MockOracle.ts 	new file:   blockchain/typechain/contracts/mocks/Token.ts 	new file:   blockchain/typechain/contracts/mocks/index.ts 	modified:   blockchain/typechain/factories/@chainlink/contracts/src/v0.8/index.ts 	modified:   blockchain/typechain/factories/@chainlink/contracts/src/v0.8/interfaces/index.ts 	new file:   blockchain/typechain/factories/@openzeppelin/contracts/token/ERC721/extensions/ERC721Enumerable__factory.ts 	new file:   blockchain/typechain/factories/@openzeppelin/contracts/token/ERC721/extensions/IERC721Enumerable__factory.ts 	modified:   blockchain/typechain/factories/contracts/AminoChainAuthenticator__factory.ts 	new file:   blockchain/typechain/factories/contracts/AminoChainMarketplace__factory.ts 	new file:   blockchain/typechain/factories/contracts/DonationNFT__factory.ts 	modified:   blockchain/typechain/factories/contracts/index.ts 	new file:   blockchain/typechain/factories/contracts/interfaces/IAminoChainAuthenticator__factory.ts 	new file:   blockchain/typechain/factories/contracts/interfaces/IAminoChainMarketplace__factory.ts 	new file:   blockchain/typechain/factories/contracts/interfaces/IDonationNFT__factory.ts 	new file:   blockchain/typechain/factories/contracts/interfaces/index.ts 	new file:   blockchain/typechain/factories/contracts/mocks/MockAminoChainMarketplace__factory.ts 	new file:   blockchain/typechain/factories/contracts/mocks/MockERC20__factory.ts 	new file:   blockchain/typechain/factories/contracts/mocks/MockNFT copy.sol/MockNFT__factory.ts 	new file:   blockchain/typechain/factories/contracts/mocks/MockNFT copy.sol/index.ts 	new file:   blockchain/typechain/factories/contracts/mocks/MockNFT__factory.ts 	new file:   blockchain/typechain/factories/contracts/mocks/MockOracle__factory.ts 	new file:   blockchain/typechain/factories/contracts/mocks/Token__factory.ts 	new file:   blockchain/typechain/factories/contracts/mocks/index.ts 	modified:   blockchain/typechain/index.ts 	new file:   blockchain/utils/verify.ts 	modified:   blockchain/yarn.lock

+++ blockchain/contracts/AminoChainAuthenticator.sol
@@ -1,38 +1,49 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.7;
 
-import "./IAminoChainAuthenticator.sol";
+import "./interfaces/IAminoChainAuthenticator.sol";
 import "./DonationNFT.sol";
-import "./IAminoChainMarketplace.sol";
+import "./interfaces/IAminoChainMarketplace.sol";
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol";
 
 contract AminoChainAuthenticator is IAminoChainAuthenticator, IERC721Receiver {
-    uint DEFAULT_PRICE = 40000*10e18;
+    uint256 DEFAULT_PRICE = 40000 * 10e18;
 
     IDonationNFT immutable nft;
     IAminoChainMarketplace immutable marketplace;
     IERC20 immutable usdc;
 
-    constructor(address nftAddress, address marketplaceAddress, address usdcAddress) {
+    constructor(
+        address nftAddress,
+        address marketplaceAddress,
+        address usdcAddress
+    ) {
         nft = IDonationNFT(nftAddress);
         nft.setApprovalForAll(marketplaceAddress, true);
         marketplace = IAminoChainMarketplace(marketplaceAddress);
         usdc = IERC20(usdcAddress);
     }
 
-    function registerUser(AminoChainLibrary.BioData calldata bioData, address biobankAddress) public {
-        uint tokenId = nft.mint(msg.sender, bioData);
+    function registerUser(AminoChainLibrary.BioData calldata bioData, address biobankAddress)
+        public
+    {
+        uint256 tokenId = nft.mint(msg.sender, bioData);
         nft.transferFrom(address(this), address(marketplace), tokenId);
         marketplace.listItem(tokenId, DEFAULT_PRICE, msg.sender, biobankAddress);
         emit UserRegistered(msg.sender);
     }
 
-    function isRegistered() public view returns(bool) {
+    function isRegistered() public view returns (bool) {
         return nft.getTokenIdByDonor(msg.sender) != 0; // fixme
     }
 
-    function onERC721Received(address, address, uint256, bytes calldata) external pure returns (bytes4) {
+    function onERC721Received(
+        address,
+        address,
+        uint256,
+        bytes calldata
+    ) external pure returns (bytes4) {
         return IERC721Receiver.onERC721Received.selector;
     }
 }
